placed unused files in own folder
