+++ academy/__manifest__.py
@@ -25,6 +25,7 @@
         "views/templates.xml",
         "views/instructor.xml",
         "views/reports.xml",
+        "demo/demo.xml",
     ],
     # only loaded in demonstration mode
     "demo": ["demo/demo.xml",],
