+++ academy/models/__init__.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from . import academy
-from . import student
 from . import session
 from . import major
 from . import instructor
 from . import wizard
+from . import course
