+++ academy/models/disconnected_models/academy.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
-
+"""
+This python model was created during the building of my first test module. Saving it for possible reference on building other modules
+"""
 from odoo import models, fields, api
 import random
